Alter some keybinds and style hyprlock input text
Also in this commit I noticed a thing (not sure if it's a bug or
a feature) where when you change focus in Hyprland, the window
that you used to be on animates a red border in the bottom left.
Pretty sure this wasn't always a thing and (AFAIK) I haven't updated
Hyprland so not sure what could've caused it. Putting this note here
in case I come back later and try to work out what happened.