System update
Not using latest kernel version because of issues with support for
graphics drivers on apple silicon. This ended up being fixed the day
after but I really don't want to recompile a bunch of stuff so I'm
leaving it until the next update.

List of other changes made while I was at it:
- Switched back to MacOS cursors (Bibata ones had the wrong hitboxes)
- Switched back to Adwaita dark (Tokyonight GTK theme not working anymore after update)
- Adjusted some positioning in Hyprlock config after update broke some stuff
- Added lazygit